fix: connect event

+++ furui-ebpf/src/ingress_icmp.rs
@@ -1,4 +1,4 @@
-use aya_bpf::cty::c_ushort;
+use aya_bpf::cty::c_long;
 use aya_bpf::{macros::classifier, programs::SkBuffContext};
 
 #[classifier(name = "ingress_icmp")]
@@ -9,6 +9,6 @@ pub fn ingress_icmp(ctx: SkBuffContext) -> i32 {
     }
 }
 
-unsafe fn try_ingress_icmp(_ctx: SkBuffContext) -> Result<i32, c_ushort> {
+unsafe fn try_ingress_icmp(_ctx: SkBuffContext) -> Result<i32, c_long> {
     Ok(0)
 }
